Add WKD for staff public key discovery
Web Key Directory is a HTTPS based key-discovery scheme. It allows mail
clients to look at a well-known directory on the senders domain to
automatically find and download the matching public key for a message.

This commit adds the directory structure, and future PRs can use this
structure to insert their own pubkeys.