Update our public policies according to the meeting, mainly with regards to primary / secondary contacts (#107)
* Udpate policies: remove primary / secondary GC

These are a relic from the past and more confusing than helpful.
We can still offer them for projects and communities that need them,
but in general let us go with the default and else offer tailored
solutions discussed with the project or community in question.

* Update community cloak policies

These were already decided in the last meeting and communicated,
we just forgot to update this part of the documentation. Oops.