Show user-facing help for alis (#183)
Previously, the help text for `/msg alis help list` was copied out of an opered connection and therefore mentioned oper-only options (`-showsecret`). Replace with the most recent version of the user-facing help. Some spacing changes seem to have been made in the alis output; these were maintained as well in this PR.