Edit ERC guide for brevity and make CI happy
The SASL guides are "how to set up SASL on your client" and not "how to
set up your client from scratch" -- we assume people have a working
client and just need to enable SASL on it. As such, eliminate much of
the wording in the ERC guide that has absolutely nothing to do with
SASL.

The ERC guide and a handful of other pages were modified in order to
conform to the style guide and fix up some links that caused
redirections (and therefore htmlproofer fails).