Update privacy policy
periclase will be deployed soon, which collects (but currently does not store) CTCP VERSION information from clients. Explicitly mention this collection in the privacy policy. Additionally, ensure the abuse protection purpose clearly states that information is processed to protect the staff, users, and services (as opposed to some subset of those things).