Automatically update privacy policy last modified
This is a bit of a hacky workaround, but it seemed better than adding a
dependency to the seemingly-abandoned jekyll-last-modified-at gem. This
will store an additional configuration variable containing the last
modified date in YYYY-mm-dd format to _config.yml, so that jekyll can
pull that and render it in the page.

Additionally, the full privacy policy history is now linked adjacent to
the last modified date, so people can easily check on what has changed
with the policy.