Ignore perpetually broken URL checks
liberapay rejects the link liveness check with a 403. While we could spoof the User-Agent to bypass this, it seems better to respect their wishes to block these automated processes and just not bother checking the link in the first place.

Not sure what the issue is with imy.se, but ignore that too as it seems to be an issue with the TLS handshake (?) despite the link working fine in a browser.