Point to SASL documentation from Emacs upstream
ERC for Emacs has SASL support as of 5.5. 

User may update via M-x package-list packages RET.  Then mark ERC for an update in the package list ("i" to mark), and then press "x" to begin the installation.

Please have a look at the link this PR adds for more information.