Update jekyll-build.yml
For some reason, the auto-date for privacy policy last updated was pulling the overall date the site was last updated rather than for just the privacy policy specifically. Get rid of the -- in hopes that either works or at least yields a more informative error message as to what is happening.