One final test
It seems that it's escaping the end summary tag for some reason; let's
split that up into multiple lines to see if that makes the parser
realize we want/need that tag. If this doesn't work, I have one other
thing to try, otherwise I'll revert the recent change to parse block
level markdown inside of block HTML by default.